refactor(webauthn): share the origin allowlist between both gates
SonarCloud flagged 11.6% duplication on the PR: the main process and the preload relay each carried their own copy of the origin parsing. Both now build the set from app/webauthn/originAllowlist.js.
